third_party/WebKit/LayoutTests/inspector/sass/test-ast-editing-1.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script src="./sass-test.js"></script>
<script src="./editing-test-suite.js"></script>
<style>
pre {
    font-family: monospace;
}
</style>
<script>

function test()
{
    InspectorTest.evaluateInPagePromise("getCSS()")
        .then(result => InspectorTest.parseCSS("", result.value))
        .then(ast => InspectorTest.runEditingTests(ast))
        .catch(console.error.bind(console))
        .then(InspectorTest.completeTest);
}

function getCSS()
{
    return document.querySelector(".snippet").textContent;
}

</script>
</head>

<body onload="runTest()">
<p>
Verifies AST editing of formatted source.
</p>
<pre class="snippet">
div {
    /* This is a regular comment */
    color: red;
    /* position: absolute; */
    display: block
}
</pre>
</body>
</html>
